generator/requirements.txt
dnspython==2.3.0
pymongo==4.4.0
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
